remove old docker scripts and files; move bashrc to docker subdir
The old docker scripts and files were created for doing development on
this repo, before I had ever looked at riff-rtc (then rhythm-rtc),
before the riff-docker repo existed.

A common design is to have a Dockerfile for building a container for the
project at the root of the repo. This allows it to be modified as the
project changes and you can get a Dockerfile that matches the version
checked out.

I believe 2 dockerfiles will be needed, one for production and one for
development. Splitting in this way is still to come.